chore: remove old foyer-memory code
